% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\title{Constants frequently used in asteroseismology}
\format{
A data frame containing 9 constants:
\describe{
\tabular{rl}{
\strong{a_rad} \tab radiation constant \cr
\strong{gk} \tab Gravitational constant as defined in ESTA working group \cr
\strong{pi4} \tab 4 times pi \cr
\strong{smass} \tab solar mass in cgs \cr
\strong{sradius} \tab solar radius in cgs \cr
\strong{slum} \tab solar luminosity (in cgs) \cr
\strong{stefbol} .\tab Stefan-Bolzman constant (4\emph{a}c/3) \cr
\strong{srho} \tab solar meand density (in cgs) \cr
\strong{sdynt} \tab solar dynamical time
}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
constants
}
\description{
A dataset containing the most frequent constant used in asteroseismology
and stellar science.
}
\keyword{datasets}
