% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_range.R
\name{l_range}
\alias{l_range}
\title{l_range}
\usage{
l_range(model.df)
}
\arguments{
\item{model.df}{DataFrame.}
}
\value{
It returns a vector with c(l_min, l_max)
}
\description{
\code{l_range} determines the common L range for all the
oscillation spectrum as
n = (L_min, L_max)
}
\details{
This function takes one argument, a dataframe with the synthetic oscillation
spectra read from read_graco, read_adipls, read_gyre, or read_losc.
}
\examples{
range <- l_range(df_oscillations)

\dontrun{
  l_range(df_oscillations)
}
}
\seealso{
\code{\link{n_range}}
}
