% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_sep.R
\name{l_sep}
\alias{l_sep}
\title{l_sep}
\usage{
l_sep(model.df, L)
}
\arguments{
\item{model.df}{DataFrame}

\item{L}{Scalar. Mode degree for which the large separation is calculated}
}
\value{
A vector containing the large separation for the given L
}
\description{
\code{l_sep} calculates the large frequency separation
}
\details{
This function takes one argument, a dataframe with the synthetic oscillation
spectra read from read_graco, read_adipls, read_gyre, or read_losc, and
calculates the large spacing defined as
ls = nu_b - nu_a
where
nu_b = nu_{n,l}; nu_a = nu_{n-1,l}
}
\examples{
\dontrun{
   my_ls <- l_sep(df_oscillations, 0) #This calculates the ls for radial
   modes
   }
\dontrun{
  l_sep(df_olscillations, 0)
  }

}
\seealso{
\code{\link{s_sep}}
}
