% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalz.R
\name{metalz}
\alias{metalz}
\title{Metallicity conversor}
\usage{
metalz(metallicity, z = FALSE, ypr = NULL, dydz = NULL, szx = NULL, prec = 8)
}
\arguments{
\item{metallicity}{Numeric, scalar. Input metallicty value (default Fe/H).}

\item{z}{Logic, scalar. If TRUE then metallicity is provided as Z (default F).}

\item{ypr}{Numeric, scalar. Primordial Helium. If not provided,
default is 0.26627.}

\item{dydz}{Numeric, scalar. Dy/Dz value. IF not provided,
default is 2.2}

\item{szx}{Numeric, scalar. Solar Z/H value. IF not provided,
default is  0.0245}

\item{prec}{Numeric, scalar. Required output precision (default input
precision). It indicates the number of decimals in output.}
}
\value{
A numeric vector containing c(X, Y, Z, Fe/H).  If Fe/H is provided
then Z will be returned. Likewise if Z is provided then Fe/H will be
returned.
}
\description{
\code{metalz} returns the conversion from Fe/H to Z and viceversa
}
\details{
This function requires at least one argument to work, either the metallicity
as Fe/H or as Z. Avoid to modify default values for primordial helium or
variation of Y or X with Z (see details in parameter description).
}
\examples{
metalz(feh = 0.0)
metalz(z = 0.02)
metalz(feh = 0.0, ypr = 0.26, prec = 4)

\dontrun{
metalz(feh = 0.25, z=0.02)
}
}
