% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_maxmin.R
\name{n_maxmin}
\alias{n_maxmin}
\title{n_maxmin}
\usage{
n_maxmin(model.df)
}
\arguments{
\item{model.df}{DataFrame.}
}
\value{
It returns a single numeric value
}
\description{
\code{n_maxmin} determines the largest radial order (n) among all the
min(n, L)
}
\details{
This function takes one argument, a dataframe with the synthetic oscillation
spectra read from read_graco, read_adipls, read_gyre, or read_losc.
}
\examples{
n_min <- n_maxmin(df_oscillations)

\dontrun{
  n_maxmin(df_olscillations)
}
}
