% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_minmax.R
\name{n_minmax}
\alias{n_minmax}
\title{n_minmax}
\usage{
n_minmax(model.df)
}
\arguments{
\item{model.df}{DataFrame.}
}
\value{
It returns a single numeric value
}
\description{
\code{n_minmax} determines the smallest radial order (n) among all the
max(n, L)
}
\details{
This function takes one argument, a dataframe with the synthetic oscillation
spectra read from read_graco, read_adipls, read_gyre, or read_losc.
}
\examples{
n_max <- n_minmax(df_oscillations)

\dontrun{
  n_minmax(df_olscillations)
}
}
