% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_range.R
\name{n_range}
\alias{n_range}
\title{n_range}
\usage{
n_range(model.df, mono = F)
}
\arguments{
\item{model.df}{dataframe}

\item{mono}{logical option (default FALSE) to calculate minimum n value
from which the n vector is monotonic (i.e. diff(n)=1).}
}
\value{
It returns a vector with c(n_maxmin, n_minmax)
}
\description{
\code{n_range} determines the common n range for all the L in the
oscillation spectrum as
n = (\code{\link{n_maxmin}}, \code{\link{n_minmax}})
}
\details{
This function takes one argument, a dataframe with the synthetic oscillation
spectra read from read_graco, read_adipls, read_gyre, or read_losc.
}
\examples{
range <- n_nrange(df_oscillations)

\dontrun{
  n_range(df_olscillations)
}
}
\seealso{
\code{\link{n_maxmin}}, \code{\link{n_minmax}}
}
