% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_range.R
\name{nl_range}
\alias{nl_range}
\title{nl_range}
\usage{
nl_range(model.df, mono = F)
}
\arguments{
\item{model.df}{dataframe}

\item{mono}{logical option (default FALSE) to calculate minimum n value
from which the n vector is monotonic (i.e. diff(n)=1).}
}
\value{
It returns a vector with c(l_min, l_max, n_min, n_max)
}
\description{
\code{l_range} determines the common L range for all the
oscillation spectrum as
nl_range = (n_min, n_max) & (L_min, L_max)
}
\details{
This function takes one argument, a dataframe with the synthetic oscillation
spectra read from read_graco, read_adipls, read_gyre, or read_losc.
}
\examples{
range <- nl_range(df_oscillations)

\dontrun{
  nl_range(df_oscillations)
}
}
\seealso{
\code{\link{n_range}}, \code{\link{l_range}}
}
