% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adipls.R
\name{read_adipls}
\alias{read_adipls}
\title{read_adipls}
\usage{
read_adipls(filename = NULL, header = FALSE, verbose = FALSE)
}
\arguments{
\item{filename}{String, scalar. Input ADIPLS oscillation file.}

\item{header}{Logic. If TRUE then header of file is read (default F).}

\item{verbose}{Logic. If TRUE then}
}
\value{
A column-named dataframe containing the following data
l: the mode degree
n: the mode radial node
w2: squared oscillation frequency normalized to GMR
sig: oscillation frequency in microHz
E: energy of the mode
NP: number of nodes as P mode
NG: number of nodes as G mode
}
\description{
\code{read_adipls} reads ADIPLS oscillation files
}
\details{
This function takes one argument, the ADIPLS file (.dat) to be read with full
path. It stores the result in an R dataframe.
}
\examples{
myfreqs <- read_adipls('/myhome/gracofile.dat')

\dontrun{
  read_graco('/myhome/gracofile.dat')
}
}
\seealso{
\code{\link{read_graco}}, \code{\link{read_losc}}, and
\code{\link{read_gyre}}
}
