% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cestam.R
\name{read_cestam}
\alias{read_cestam}
\title{read_cestam}
\usage{
read_cestam(strFile)
}
\arguments{
\item{file}{ASCII file (.osc or .OSC)}
}
\value{
A dataframe with the stellar structure, in which global parameters
\itemize{
\item common values for the star - are saved as attributes, and variables -
physical quantities that vary with the stellar depth- are saved as columns.
}
}
\description{
\code{read_cestam} reads the content of stellar structure files form CESTAM
code, and save it to a dataframe.
}
\details{
This function takes one argument: the stellar structure file .osc from
CESTAM code. It should also compatible with last versions (deprecated)
of CESAM2K models as well as MESA (OSC-type) models
}
\examples{
mystruc <- read_cestam(/my/path/file.osc)

\dontrun{
  mystruc <- read_cestam(/my/path/file.osc)
}
}
\seealso{
\code{\link{read_gong}}
}
