% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filou.R
\name{read_filou}
\alias{read_filou}
\title{read_filou}
\usage{
read_filou(filename = NULL, header = FALSE)
}
\arguments{
\item{filename}{String, scalar. Input FILOU oscillation file (.frq)}

\item{header}{Logic. If TRUE then header of file is read (default F)}
}
\value{
A column-named dataframe containing the following data
l: the mode degree
n: the mode radial node
w2: squared oscillation frequency normalized to GMR
sig: oscillation frequency in microHz
fc: oscillation period in days
No: oscillation constant
pc: number of nodes as P mode
I: number of nodes as G mode

\if{html}{\out{<div class="sourceCode">}}\preformatted{    All data included in file header is stored as attributes. To see all
    attributes use:
           attributes(yourobject)
    To select one particular attribute use
           attr(yourobject, "name-of-attribute")
}\if{html}{\out{</div>}}
}
\description{
\code{read_filou} reads FILOU oscillation files
}
\details{
This function takes one argument, the GraCo file (.dat) to be read with full
path. It stores the result in an R dataframe.
}
\examples{
myfreqs <- read_filou('/myhome/filoufile.frq')

\dontrun{
  read_filou('/myhome/filoufile.frq')
}

}
\seealso{
\code{\link{read_adipls}}, \code{\link{read_losc}},
\code{\link{read_graco}}, and \code{\link{read_gyre}}
}
