% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graco.R
\name{read_graco}
\alias{read_graco}
\title{read_graco}
\usage{
read_graco(filename = NULL, header = FALSE)
}
\arguments{
\item{filename}{String, scalar. Input GraCo oscillation file}

\item{header}{Logic. If TRUE then header of file is read (default F)}
}
\value{
A column-named dataframe containing the following data
        l: the mode degree
        n: the mode radial node
        w2: squared oscillation frequency normalized to GMR
        sig: oscillation frequency in microHz
        P: oscillation period in days
        Q: oscillation constant
        NP: number of nodes as P mode
        NG: number of nodes as G mode
}
\description{
\code{read_graco} reads GraCo oscillation files
}
\details{
This function takes one argument, the GraCo file (.dat) to be read with full
path. It stores the result in an R dataframe.

\usage{
  read_graco(filename, header = F)
}
}
\examples{
myfreqs <- read_graco('/myhome/gracofile.dat')

\dontrun{
  read_graco('/myhome/gracofile.dat')
}
}
