% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gyre.R
\name{read_gyre}
\alias{read_gyre}
\title{read_gyre}
\usage{
read_gyre(filename = NULL, header = FALSE)
}
\arguments{
\item{filename}{String, scalar. GYRE oscillation file.}

\item{header}{Logic. If TRUE then header of file is read (default F).}
}
\value{
A column-named dataframe containing the following data
l: the mode degree
npg: the mode radial node
np: number of nodes as P mode
ng: onumber of nodes as G mode
sig: frequency in microHz
sigi: imaginary part of the frequency
}
\description{
\code{read_gyre} reads GYRE oscillation files
}
\details{
This function takes one argument, the GYRE file (.txt) to be read with full
path. It stores the result in an R dataframe.
}
\examples{
myfreqs <- read_gyre('/myhome/gyrefile.txt')

\dontrun{
  read_gyre('/myhome/gyrefile.txt')
}
}
\seealso{
\code{\link{read_gyre}}, \code{\link{read_losc}}, and
\code{\link{read_gyre}}
}
