% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_losc.R
\name{read_losc}
\alias{read_losc}
\title{read_losc}
\usage{
read_losc(filename = NULL, header = FALSE)
}
\arguments{
\item{filename}{String, scalar. Input LOSC oscillation file}

\item{header}{Logic. If TRUE then header of file is read (default F)}
}
\value{
A column-named dataframe containing the following data
        n: the mode radial node
        l: the mode degree
        sig: oscillation frequency in Hz
}
\description{
\code{read_losc} reads GraCo oscillation files
}
\details{
This function takes one argument, the LOSC file (.txt) to be read with full
path. It stores the result in an R dataframe.

\usage{
  read_losc(filename, header = F)
}
}
\examples{
myfreqs <- read_losc('/myhome/loscfile.dat')

\dontrun{
  read_losc('/myhome/loscfile.dat')
}
}
