% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readobs_corot.R
\name{readobs_corot}
\alias{readobs_corot}
\title{readobs_corot}
\usage{
readobs_corot(corotfits, hdu = 1)
}
\arguments{
\item{hdu}{Numeric, scalar. Data block: 1 (default) or 2,3 (additional data)}

\item{filename}{String, scalar. Input CoRoT FITS  file (.FITS)}
}
\value{
A column-named dataframe containing the lightcurve and related data
as well as the original header as attributes.

Header data is stored as attributes. To see all
attributes use:
\code{attributes(yourobject)}
To select one particular attribute use
\code{attr(yourobject, "name-of-attribute")}
}
\description{
\code{readobs_corot} reads CoRoT FITS files
}
\details{
This function reads CoRoT mission FITS files and stores the result in
an R dataframe. The function reads
\itemize{
\item AN2_STAR files. Former Seismo Field. These files correspond to bright stars
observed by CoRoT (V<=8). It contains two binary blocks accessed by the hdu
parameter.
\item EN2_STAR_CH files. Former Exo Field. These files correspond to faint stars
observed by CoRoT (8<V<12). It contains three binary blocks accessed by the
hdu parameter.
}

Details about the content of each file are given below. All the information
has been extracted from the publication
\link{CoRoT Legacy Book}(https://www.edp-open.org/books/edp-open-books/320-the-
corot-legacy-book)
\subsection{AN2_STAR (hdu=1)}{
\itemize{
\item DATETT: (double) date of measurement in Terrestrial Time; terrestrial
time, Julian date
\item RAWFLUX: (float) Integrated red flux; e- per 32s (32-s or 512-s sampling)
\item RAWFLUXDEV: (float) Standard deviation of the 16 exposures of 32 s in
the green channel added for the 512 sampling; e- per 32s
\item RAWSTATUS: (int)
\item BG: (float) Background flux already subtracted; e- per 32s
}
}

\subsection{AN2_STAR (hdu=2)}{
\itemize{
\item DATEBARTT: (double) date of the measurement in the solar barycentric
reference frame; Solar barycentric Terrestrial Time, Julian day
\item FLUXBAR: (float) White flux, after the gap filling correction;
e- per 32s
\item FLUXDEVBAR: (float)
\item STATUSBAR: (int) Flag of the status
}
}

\subsection{AN2_STAR (hdu=3)}{
\itemize{
\item DATEBARREGTT: (double) date of the measurement in the solar barycentric
reference frame; Solar barycentric Terrestrial Time, Julian day.
\item FLUXBARREG: (float) White flux, after the gap filling correction;
e- per 32s
\item FLUXDEVBARREG: (float)
\item STATUSBARREG: (int) Flag of the status
}
}

\subsection{EN2_STAR_CH  (hdu=1)}{
\itemize{
\item DATE: (string) calendar date, yyyy-mm-ddThh:mm:ss
\item DATETT: (double) date of measurement in Terrestrial Time; terrestrial
time, Julian date
\item DATEBARTT: (double ) date of the measurement in the solar barycentric
reference frame; Solar barycentric Terrestrial Time, Julian day
\item STATUS: (int) Flag of the status
\item REDFLUX: (float) Integrated red flux; e- per 32s (32s or 512s sampling)
\item REDFLUXDEV: (float) Standard deviation of the 16 exposures of 32 s in
the green channel added for the 512 sampling; e- per 32s
\item GREENFLUX: (float) Integrated green flux; e- per 32s (32s or 512s sampling)
\item GREENFLUXDEV: (float) Standard deviation of the 16 exposures of 32 s in
the green channel added for the 512 sampling; e- per 32s
\item BLUEFLUX: (float) Integrated blue flux; e- per 32s (32s or 512s sampling)
\item BLUEFLUXDEV: (float) Standard deviation of the 16 exposures of 32 s in
the green channel added for the 512 sampling; e- per 32s
\item WHITEFLUX: (float) Integrated white flux; e- per 32s
\item BG: (float) Background flux already subtracted; e- per 32s
}
}

\subsection{EN2_STAR_CH (hdu=2)}{
\itemize{
\item DATETT: (double) date of measurement in Terrestrial Time; terrestrial
time, Julian date
\item DATEBARTT: (double ) date of the measurement in the solar barycentric
reference frame; Solar barycentric Terrestrial Time, Julian day
\item WHITEFLUXFIL: (float) White flux, after the gap filling correction;
e- per 32s
\item STATUSFIL: (int) Flag of the status
\item TEXP: (int) Exposure Time; 32s or 512s
}
}
}
\examples{
myLC_exo <- readobs_corot('/myhome/EN2_STAR_CHR_corotid_tini-tfin.fits')
myLC_sis <- readobs_corot('/myhome/AN2_STAR_corotid_tini-tfin.fits')

\dontrun{
  readobs_corot('/myhome/EN2_STAR_CHR_corotid_tini-tfin.fits')
}

}
