% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_sep.R
\name{s_sep}
\alias{s_sep}
\title{s_sep}
\usage{
s_sep(model.df, L)
}
\arguments{
\item{model.df}{DataFrame}

\item{L}{Scalar. Lowest degree for which the large separation is calculated}
}
\value{
A vector containing the large separation for the given L
}
\description{
\code{s_sep} calculates the small frequency separation
}
\details{
This function takes one argument, a dataframe with the synthetic oscillation
spectra read from read_graco, read_adipls, read_gyre, or read_losc, and
calculates the small spacing defined as\preformatted{ss = nu_b - nu_a
}

where
nu_b = nu_{n,l}; nu_a = nu_{n-1,l+2}

Note that the provided L is the lowest, so the dataframe must contain the
frequencies corresponding to L+2 modes.
}
\examples{
\dontrun{
 my_ss <- l_sep(df_oscillations, 0) #This calculates the ls for radial modes
}
\dontrun{
  s_sep(df_olscillations, 0)
}

}
\seealso{
\code{\link{l_sep}}
}
