% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echelled.R
\name{echelled}
\alias{echelled}
\title{read_filou}
\usage{
echelled(df_ospec = NULL, per = NULL, collapse = F)
}
\arguments{
\item{df_ospec}{Dataframe. It must contain at least a column (after
the name of sig) that contains the oscillation frequencies}

\item{per}{Numeric. The value of the periodic pattern with which the
frequencies will be sliced.}
}
\value{
Dataframe. The input data frame with an additional column with the
frequencies scaled by per:
sig: (double float) oscillation frequency (frequency units)
ech: (double float) sig modulo per (frequency units)
}
\description{
\code{echelled} computes echelle diagram from observed oscillation spectrum
}
\details{
This function takes two arguments, a dataframe containing observed
frequencies and the periodicity with which the frequencies are to be scaled
as
frequency (modulo) periodicity
}
\seealso{
\code{\link{read_adipls}}, \code{\link{read_losc}},
\code{\link{read_graco}}, and \code{\link{read_gyre}}
}
