% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_sep.R
\name{l_sep}
\alias{l_sep}
\title{l_sep}
\usage{
l_sep(model.df, L = NULL, n_inf = NULL, n_sup = NULL, vector = FALSE)
}
\arguments{
\item{model.df}{DataFrame}

\item{L}{Scalar. Mode degree for which the large separation is calculated}

\item{n_inf}{Scalar. Minimum radial order (n) considered. If NULL (default)
minimum is calculated using tostada::nrange function}

\item{n_sup}{Scalar. Maximum radial order (n) considered. If NULL (default)
maximum is calculated using tostada::nrange function}

\item{vector}{Logical. Indicates wether user want a single value or the list
of spacings for the indicated L.}
}
\value{
Large separation calculated from frequencies for the chosen L value.
If L is not provided then l_sep returns the median value from all the Dnu(L)
When n range is not specified, l_sep uses tostada::n_range function to
obtain the minimum and maximum n (common to all L values). Otherwise the
Dnu value is calculated within the range (n_min, n_max).
When vector = T l_sep gives a vector with
\itemize{
\item If no L specified: \link{Dnu(L=0), Dnu(L=1), ...(Dnu(L=Lmax)}
\item If L specified: \link{dnu_n2-n1, dnu_n3-n2, dnu_n4-n3,...dnu_n-(n-1)}
}
}
\description{
\code{l_sep} calculates the large frequency separation
}
\details{
This function takes one argument, a dataframe with the synthetic oscillation
spectra read from read_graco, read_adipls, read_gyre, or read_losc, and
calculates the large spacing defined as the vector
dnu = nu_b - nu_a
where
nu_b = nu_{n,l}; nu_a = nu_{n-1,l}
Single value result is calculated as
ls = median(dnu)
When all L are considered then
ls = median(median(dnu,L=0), median(dnu,L=1), ...)
}
\examples{
\dontrun{
   my_ls <- l_sep(df_oscillations, L=0, n_inf = 2, n_sup =10) # This
   calculates the median of the ls for radial modes
   }
\dontrun{
  my_ls <- l_sep(df_oscillations, L=0, vector = T) # This
   calculates the ls for radial modes, and gives a vector with values.
  }

}
\seealso{
\code{\link{s_sep}}
}
