% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filou.R
\name{read_filou}
\alias{read_filou}
\title{read_filou}
\usage{
read_filou(filename = NULL)
}
\arguments{
\item{filename}{String, scalar. Input FILOU oscillation file (.frq)}
}
\value{
A column-named dataframe containing the following data
l: (integer) mode degree
n: (integer) mode radial node
w2: (double float) squared oscillation frequency normalized to GMR
sig: (double float) oscillation frequency in microHz
fc: (integer) Id of the mode with which is coupled
No: Mode Id
pc: (double float) coupling coefficient
I: (double float) mode inertia

\if{html}{\out{<div class="sourceCode">}}\preformatted{    All data included in file header is stored as attributes. To see all
    attributes use:
           attributes(yourobject)
    To select one particular attribute use
           attr(yourobject, "name-of-attribute")
}\if{html}{\out{</div>}}
}
\description{
\code{read_filou} reads FILOU oscillation files
}
\details{
This function takes one argument, the GraCo file (.dat) to be read with full
path. It stores the result in an R dataframe.
}
\examples{
myfreqs <- read_filou('/myhome/filoufile.frq')

\dontrun{
  read_filou('/myhome/filoufile.frq')
}

}
\seealso{
\code{\link{read_adipls}}, \code{\link{read_losc}},
\code{\link{read_graco}}, and \code{\link{read_gyre}}
}
